feat: improve url2path mapping
Add lenght constraint into config so it can be tweaked as needed, for
example for shorter format which is required on Kaggle since 255, which
is limit on most file systems, is too long.

Cover case when there are multiple URLs that produce the same filename
when cropped. To avoid this, a random number is appended.

Change signature of `to_filename` to make it the same as
`url2file_path`'s to avoid possible confusion and mistakes.