feat: improve origin sublink filtering
Filter sublinks of origin page using common URL format instead
of matching part before '?' and '=', or before last '/'. For
this purpose, _CrawlManager__origin_candidates is changed so
it containts children URLs instead of the count. Those URLs
are used for deriving common format. Result value is also
changed to contain directory so the format can be passed to
the caller.

Another filter is also introduced, stucture link filtering. Those
links are the links that come from table, ol or ul tags. It is set
so only links that are located somewhere in the hierarcy of
mentioned tags are considered as profile page tags.