refactor: rearrange and simplify code in main.py
Remove extra statement for `scraping` initialization.

Initialize top-lvl `constants` with an object instead of a ref.
Rearrange code so it can be easily understood. This includes:
breaking down main func and extracting new function for
input gen, moving some lines to another part in the block, etc.
Remove unnecessary blank lines. Add comments to clarify what is
going on in that part of the code.